Cleaned PPSGeometryBuilder and add geader for PPSGeometryESProducer
