+++ CondTools/Geometry/plugins/BuildFile.xml
@@ -1,32 +1,31 @@
-<use name="root"/>
-<use name="dd4hep"/>
-<use name="CondCore/DBOutputService"/>
-<use name="CondFormats/Common"/>
-<use name="CondFormats/GeometryObjects"/>
-<use name="DataFormats/CTPPSDetId"/>
-<!--<use name="DetectorDescription/Core"/>-->
-<use name="DetectorDescription/DDCMS"/>
-<use name="FWCore/Framework"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/ServiceRegistry"/>
-<use name="Geometry/Records"/>
-<use name="Geometry/TrackerGeometryBuilder"/>
-<use name="Geometry/TrackerNumberingBuilder"/>
-<use name="Geometry/CaloGeometry"/>
-<use name="Geometry/EcalAlgo"/>
-<use name="Geometry/HcalCommonData"/>
-<use name="Geometry/HcalTowerAlgo"/>
-<use name="Geometry/ForwardGeometry"/>
-<use name="Geometry/CSCGeometryBuilder"/>
-<use name="Geometry/DTGeometry"/>
-<use name="Geometry/DTGeometryBuilder"/>
-<use name="Geometry/RPCGeometry"/>
-<use name="Geometry/RPCGeometryBuilder"/>
-<use name="Geometry/GEMGeometry"/>
-<use name="Geometry/GEMGeometryBuilder"/>
-<use name="Geometry/MuonNumbering"/>
-<use name="Geometry/HGCalGeometry"/>
-<use name="Geometry/VeryForwardGeometryBuilder"/>
 <library file="*.cc" name="CondToolsGeometryPlugins">
+  <use name="root"/>
+  <use name="dd4hep"/>
+  <use name="CondCore/DBOutputService"/>
+  <use name="CondFormats/Common"/>
+  <use name="CondFormats/GeometryObjects"/>
+  <use name="DataFormats/CTPPSDetId"/>
+  <use name="DetectorDescription/DDCMS"/>
+  <use name="FWCore/Framework"/>
+  <use name="FWCore/ParameterSet"/>
+  <use name="FWCore/ServiceRegistry"/>
+  <use name="Geometry/Records"/>
+  <use name="Geometry/TrackerGeometryBuilder"/>
+  <use name="Geometry/TrackerNumberingBuilder"/>
+  <use name="Geometry/CaloGeometry"/>
+  <use name="Geometry/EcalAlgo"/>
+  <use name="Geometry/HcalCommonData"/>
+  <use name="Geometry/HcalTowerAlgo"/>
+  <use name="Geometry/ForwardGeometry"/>
+  <use name="Geometry/CSCGeometryBuilder"/>
+  <use name="Geometry/DTGeometry"/>
+  <use name="Geometry/DTGeometryBuilder"/>
+  <use name="Geometry/RPCGeometry"/>
+  <use name="Geometry/RPCGeometryBuilder"/>
+  <use name="Geometry/GEMGeometry"/>
+  <use name="Geometry/GEMGeometryBuilder"/>
+  <use name="Geometry/MuonNumbering"/>
+  <use name="Geometry/HGCalGeometry"/>
+  <use name="Geometry/VeryForwardGeometryBuilder"/>
   <flags EDM_PLUGIN="1"/>
 </library>
